refactor: rename the analyzer package to weakcrypto
CI's golangci-lint (v2.7.2; local was v2.12.2, which did not flag it)
rejected 'package crypto' under revive var-naming: it shadows the standard
library.

Renamed rather than excluded. .golangci.yml already carries an exclusion
for the same complaint on plugin/, so suppression was available -- but a
package named crypto that is not stdlib crypto, inside a codebase that
imports real crypto/md5 and crypto/sha1, is genuinely confusing to read.
weakcrypto says what it is.
