+++ core/analyzers/weakcrypto/crypto_test.go
@@ -1,4 +1,4 @@
-package crypto
+package weakcrypto
 
 import (
 	"context"
